#294 ref: create RelationToDebtBalanceTrait
